+++ src/base/text_edit.h
@@ -368,6 +368,21 @@ ret_t text_edit_paint(text_edit_t* text_edit, canvas_t* c);
 ret_t text_edit_layout(text_edit_t* text_edit);
 
 /**
+ * @method text_edit_muti_line_insert_text_layout
+ * 重新排版。
+ * @param {text_edit_t*} text_edit text_edit对象。
+ * @param {uint32_t} offset 插入的字符串偏移。
+ * @param {uint32_t} insert_length 插入的字符串长度。
+ * @param {wchar_t*} wtext 插入的字符串。
+ * @param {bool_t} overwrite 是否为覆写模式。
+ * @param {uint32_t} rm_num 移除的字符串长度。
+ *
+ * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
+ */
+ret_t text_edit_muti_line_insert_text_layout(text_edit_t* text_edit, uint32_t offset, uint32_t insert_length,
+                                             const wchar_t* wtext, bool_t overwrite, uint32_t rm_num);
+
+/**
  * @method text_edit_set_offset
  * 设置滚动偏移。
  * @param {text_edit_t*} text_edit text_edit对象。